# SslSocket::shutdown logs
# VMS-5018 add multicast warning
#CLOUD-941 Added generation of cloudModules.xml based on template
#VMS-5565 Added more logs
Add language selector to lozalised files
Bootstrapper - detect language
Coding style fixes
Enabled -Werror on cloud_connectivity_ut
First two languages
Fix header styles
Fix language selector
Fix language.json
Fixed client crash when hovering over bookmarks
Fixed compile error
Fixed: wrong result code in mediator logs
Fun tests old: disable obsolete tests
Language selector
Project list file for 2.6
Senturian iphone build
Try to fix 'Bad substitution'
UT-42
Use language list from config.js
VMS-1861 Send statistic after updating system
VMS-5245: Small fix for Cellinx ptz mapping
VMS-5337: Add Rest API breaking changes info to web-admin
VMS-5493: Setup Wizard must be translatable
VMS-5512: Screenshot has to be in hi-res
VMS-5530: Export Audio in nov doesn't work
VMS-5541 - Crash detected in client: f67501de98221a023868bf0e9217e09b
VMS-5542 Axis speakers don't work
VMS-5556 Rtsp stream works in vlc not in Nx
VMS-5557 NX1 crash at stop (build 3.0.0.14114) - libc_do_syscall
VMS-5557 fix
VMS-5559 set ulimit file-max doesn't work
VMS-5567: Vista camera tilt control fix
VMS-5577 NIC can be disabled during hardware id calculation
VMS-5578: TestCamera - support separate camera-files
VMS-5587: Couldn't bound system to Cloud because /api/cloud_modules.xml is blocked by some other software
VMS-5591:Video layout fix for AV20365
VMS-5597 Hikvision DS-2CD6362F-I stream profiles are attached to the channels incorrectly
VMS-5600: License failover period should be 30 days
VMS-5601: NX1 mediaserver crash with signal 11, Segmentation fault ( FrameTypeExtractor::getFrameType)
WEB-475: 'Merge systems' header is visible for user who can't merge
WEB-477: Set port value to '1' -> the site can't be reached
generate language.json
hotfix - misspelling
ios signing for senturian
turkish translations-take 2
vms-4405. Backout.
vms-4405. cdb result code UT refactored.
vms-5571. Measure archive cleanup speed.
vms-5589. Don't select storages with free space below 150 mb for recording.
vms-5609. Removed redundant warnings.

<==== Mercurial data ====>
Branch: dev_3.0.0_gui

+++ unit_tests/cloud_connectivity_ut/src/tunnel/udp_hole_punching_acceptor_ut.cpp
@@ -196,7 +196,7 @@ protected:
             [=](SystemError::ErrorCode code, size_t size)
         {
             ASSERT_EQ(code, SystemError::noError);
-            ASSERT_EQ(buffer->size(), size);
+            ASSERT_EQ((size_t)buffer->size(), size);
             buffer->resize(0);
             socket->readSomeAsync(
                 buffer.get(),
